requirements.txt
Django==4.1.6
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
drf-spectacular
dask==2023.2.0
dask[dataframe]